Clippy fixes in integration test.
