+++ lib/tests/integration/send_event.rs
@@ -1,9 +1,7 @@
-#[cfg(test)]
-use serde_json::{json};
-use antithesis_sdk_rust::{lifecycle};
-mod common;
+use antithesis_sdk_rust::lifecycle;
+use serde_json::json;
 
-use common::{AntithesisSdk, SDKInput};
+use crate::common::{self, SDKInput};
 
 const LOCAL_OUTPUT: &str = "ANTITHESIS_SDK_LOCAL_OUTPUT";
 
@@ -14,30 +12,33 @@ const LOCAL_OUTPUT: &str = "ANTITHESIS_SDK_LOCAL_OUTPUT";
 // {"antithesis_setup":{"details":{"age":4,"name":"Tweety Bird","phones":["+1 9734970340"]},"status":"complete"}}
 // ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
 #[test]
-fn sdk_info() {
-
-    let output_file = "/tmp/antithesis-sdk.json";
+fn send_event() {
+    let output_file = "/tmp/antithesis-send-event.json";
     let prev_v = common::env::set_var(LOCAL_OUTPUT, output_file);
-    let no_details = json!({});
+    let details = json!({
+        "x": 100,
+        "tag": "last value"
+    });
 
     // only added to force the antithesis_sdk info to be generated
-    lifecycle::setup_complete(&no_details);
+    lifecycle::send_event("logging", &details);
 
     // verify the output has landed in the expected file
     match common::read_jsonl_tags(output_file) {
         Ok(x) => {
             for obj in x.iter() {
-                match obj {
-                    SDKInput::AntithesisSdk(AntithesisSdk{language, protocol_version, sdk_version: _}) => {
-                        assert_eq!(protocol_version, "1.0.0");
-                        assert_eq!(language.name, "Rust")
-                    },
-                    _ => ()
+                if let SDKInput::SendEvent {
+                    event_name,
+                    details,
+                } = obj
+                {
+                    assert_eq!(event_name, "logging");
+                    assert_eq!(&details["x"], 100);
+                    assert_eq!(&details["tag"], "last value");
                 }
             }
-        },
-        Err(e) => println!("{}", e)
+        }
+        Err(e) => println!("{}", e),
     }
     common::env::restore_var(LOCAL_OUTPUT, prev_v);
 }
-
