+++ lib/tests/integration/setup_complete_with_details.rs
@@ -1,9 +1,7 @@
-#[cfg(test)]
-use serde_json::{json};
-use antithesis_sdk_rust::{lifecycle};
-mod common;
+use antithesis_sdk_rust::lifecycle;
+use serde_json::{json, Value};
 
-use common::{SDKInput};
+use crate::common::{self, AntithesisSetup, SDKInput};
 
 const LOCAL_OUTPUT: &str = "ANTITHESIS_SDK_LOCAL_OUTPUT";
 
@@ -14,35 +12,30 @@ const LOCAL_OUTPUT: &str = "ANTITHESIS_SDK_LOCAL_OUTPUT";
 // {"antithesis_setup":{"details":{"age":4,"name":"Tweety Bird","phones":["+1 9734970340"]},"status":"complete"}}
 // ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
 #[test]
-fn send_event() {
-
-    let output_file = "/tmp/antithesis-send-event.json";
+fn setup_complete_with_details() {
+    let output_file = "/tmp/antithesis-lifecycle-with-details.json";
     let prev_v = common::env::set_var(LOCAL_OUTPUT, output_file);
-    let details = json!({
-        "x": 100,
-        "tag": "last value"
+    let bird_value: Value = json!({
+        "name": "Tweety Bird",
+        "age": 4,
+        "phones": [
+        "+1 9734970340"
+    ]
     });
-
-    // only added to force the antithesis_sdk info to be generated
-    lifecycle::send_event("logging", &details);
+    lifecycle::setup_complete(&bird_value);
 
     // verify the output has landed in the expected file
     match common::read_jsonl_tags(output_file) {
         Ok(x) => {
             for obj in x.iter() {
-                match obj {
-                    SDKInput::SendEvent{event_name, details} => {
-                        assert_eq!(event_name, "logging");
-                        assert_eq!(&details["x"], 100);
-                        assert_eq!(&details["tag"], "last value");
-                    },
-                    _ => ()
+                if let SDKInput::AntithesisSetup(AntithesisSetup { status, details }) = obj {
+                    assert_eq!(status, "complete");
+                    assert_eq!(details, &bird_value)
                 }
+                println!("{:?}", obj)
             }
-        },
-        Err(e) => println!("{}", e)
+        }
+        Err(e) => println!("{}", e),
     }
     common::env::restore_var(LOCAL_OUTPUT, prev_v);
 }
-
-
